+++ scripts/remove_fail2ban.sh
@@ -1,16 +1,16 @@
 #!/usr/bin/env bash
 
 # fail2ban Uninstaller
-# Min. Requirement  : GNU/Linux Ubuntu 16.04
-# Last Build        : 05/06/2021
+# Min. Requirement  : GNU/Linux Ubuntu 18.04
+# Last Build        : 10/12/2021
 # Author            : MasEDI.Net (me@masedi.net)
 # Since Version     : 2.1.0
 
 # Include helper functions.
-if [ "$(type -t run)" != "function" ]; then
-    BASEDIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )
+if [[ "$(type -t run)" != "function" ]]; then
+    BASE_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )
     # shellcheck disable=SC1091
-    . "${BASEDIR}/helper.sh"
+    . "${BASE_DIR}/helper.sh"
 fi
 
 # Make sure only root can run this installer script.
@@ -27,22 +27,23 @@ function init_fail2ban_removal() {
 
         run apt-get remove --purge -qq -y fail2ban
     else
-        echo "Fail2ban package not found, possibly installed from source."
+        info "Fail2ban package not found, possibly installed from source."
 
         run rm -f /usr/local/bin/fail2ban-*
     fi
 
     run dpkg --purge fail2ban
+
     [ -f /etc/systemd/system/multi-user.target.wants/fail2ban.service ] && \
         run unlink /etc/systemd/system/multi-user.target.wants/fail2ban.service
-    [ -f /lib/systemd/system/fail2ban.service ] && \
-        run rm /lib/systemd/system/fail2ban.service
+    [ -f /lib/systemd/system/fail2ban.service ] && run rm /lib/systemd/system/fail2ban.service
 
     # Remove fail2ban config files.
+    echo "Removing fail2ban configuration..."
     warning "!! This action is not reversible !!"
 
-    if "${AUTO_REMOVE}"; then
-        if [[ ${FORCE_REMOVE} == true ]]; then
+    if [[ "${AUTO_REMOVE}" == true ]]; then
+        if [[ "${FORCE_REMOVE}" == true ]]; then
             REMOVE_FAIL2BAN_CONFIG="y"
         else
             REMOVE_FAIL2BAN_CONFIG="n"
@@ -53,27 +54,28 @@ function init_fail2ban_removal() {
         done
     fi
 
-    if [[ "${REMOVE_FAIL2BAN_CONFIG}" == Y* || "${REMOVE_FAIL2BAN_CONFIG}" == y* ]]; then
+    if [[ "${REMOVE_FAIL2BAN_CONFIG}" == y* || "${REMOVE_FAIL2BAN_CONFIG}" == Y* ]]; then
         [ -d /etc/fail2ban/ ] && run rm -fr /etc/fail2ban/
 
-        echo "All your fail2ban configuration files deleted permanently."
+        echo "All configuration files deleted permanently."
     fi
 
     # Final test.
-    if "${DRYRUN}"; then
-        info "Fail2ban server removed in dryrun mode."
-    else
+    if [[ "${DRYRUN}" != true ]]; then
         if [[ -z $(command -v fail2ban-server) ]]; then
             success "Fail2ban server removed succesfully."
         else
             info "Unable to remove fail2ban server."
         fi
+    else
+        info "Fail2ban server removed in dry run mode."
     fi
 }
 
 echo "Uninstalling fail2ban server..."
+
 if [[ -n $(command -v fail2ban-server) ]]; then
-    if "${AUTO_REMOVE}"; then
+    if [[ "${AUTO_REMOVE}" == true ]]; then
         REMOVE_FAIL2BAN="y"
     else
         while [[ "${REMOVE_FAIL2BAN}" != "y" && "${REMOVE_FAIL2BAN}" != "n" ]]; do
@@ -81,7 +83,7 @@ if [[ -n $(command -v fail2ban-server) ]]; then
         done
     fi
 
-    if [[ "${REMOVE_FAIL2BAN}" == Y* || "${REMOVE_FAIL2BAN}" == y* ]]; then
+    if [[ "${REMOVE_FAIL2BAN}" == y* || "${REMOVE_FAIL2BAN}" == Y* ]]; then
         init_fail2ban_removal "$@"
     else
         echo "Found fail2ban server, but not removed."
